Fcoralsasso add tests addPeer removePeer (#257)
* Removed duplicated code on comment

* Added new unit tests for methods addPeer and removePeer

* Added new paths to be used in tests

* Fixed typo

* Added verification for peer before deleting it

* Ran flake8 and fixed warnings

+++ candig/server/datarepo.py
@@ -3406,9 +3406,10 @@ class SqlDataRepository(AbstractDataRepository):
         """
         Remove peers by URL.
         """
-        q = models.Peer.delete().where(
-            models.Peer.url == url)
-        q.execute()
+        if self.getPeer(url):
+            q = models.Peer.delete().where(
+                models.Peer.url == url)
+            q.execute()
 
     def removePeers(self):
         """
